chore: generate research-factory-issue workflow and update change-factory
